feat: implement BaseScreen and UI screen standardisation

+++ main/domains/definitions.h
@@ -35,7 +35,7 @@ static const uint8_t LCD_RGB_PCLK_ACTIVE_NEG = 1;         // 0: rising edge, 1: 
 static const uint8_t LCD_RGB_DATA_WIDTH = 16;             //  8 | 16
 static const uint8_t LCD_RGB_PIXEL_BITS = 16;             // 24 | 16
 static const uint8_t LCD_RGB_FRAME_BUF_NUM = 1;           // 1/2/3
-static const uint8_t LCD_RGB_BOUNCE_BUF_SIZE = 800 * 480; // Bounce buffer size in bytes. This function is used to avoid screen drift.
+static const uint32_t LCD_RGB_BOUNCE_BUF_SIZE = 800 * 480; // Bounce buffer size in bytes. This function is used to avoid screen drift.
                                                           // To enable the bounce buffer, set it to a non-zero value. Typically set to `ESP_PANEL_LCD_WIDTH * 10`
                                                           // The size of the Bounce Buffer must satisfy `width_of_lcd * height_of_lcd = size_of_buffer * N`,
                                                           // where N is an even number.
